% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{colselecter_text}
\alias{colselecter_text}
\title{colselecter_text}
\usage{
colselecter_text(inval)
}
\arguments{
\item{inval}{either Starting with or Ending with or Containing}
}
\description{
Convenience function to clean code below
Designed to provide a text argument for the column selections
}
