% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewenhanceAddin_calling.R
\name{datacheck}
\alias{datacheck}
\title{datacheck}
\usage{
datacheck(datain)
}
\arguments{
\item{datain}{a data frame or list}
}
\value{
A character list of data frames
}
\description{
Check whether an input has been given. If it hasn't, look in global environment
To see if there are any data frames.
}
