% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{gen_ui}
\alias{gen_ui}
\title{Generate UI for the gadget.}
\usage{
gen_ui(datain, datalist)
}
\arguments{
\item{datain}{dataframe or list}

\item{datalist}{string with names of data frames}
}
\description{
Generate UI for the gadget.
}
