% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewenhanceAddin_calling.R
\name{get_columnselect_text}
\alias{get_columnselect_text}
\title{Column text selector}
\usage{
get_columnselect_text(expression, intext, labelorname, dataname,
  contains = FALSE)
}
\arguments{
\item{expression}{what function is being checked}

\item{intext}{the text to search over}

\item{labelorname}{the input value, label or name}

\item{dataname}{name of data frame}

\item{contains}{if expression is contains, this is TRUE}
}
\description{
Handles the text string given by user for starts with/ ends with/ contains
}
