% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_in.R
\name{\%not in\%}
\alias{\%not in\%}
\title{\%not in\%}
\usage{
x \%not in\% table
}
\arguments{
\item{x}{a vector to compare}

\item{table}{list of elements we want to exclude from x}
}
\description{
This is the inverse of \%in\%
It is a convenience function for use in this package
}
