% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelgetter.R
\name{names_label}
\alias{names_label}
\title{name look}
\usage{
names_label(datain, label_name = "Name")
}
\arguments{
\item{datain}{inputted data set}

\item{label_name}{whether you want the names or labels}
}
\value{
a vector of labels or names depending on the value of label_name
}
\description{
This is a function which either gives the names of an object
or the labels, depending on the inputs given
}
