% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server_in}
\alias{server_in}
\title{Server code for the gadget.}
\usage{
server_in(input, output, session)
}
\arguments{
\item{input}{inputs to the server}

\item{output}{outputs from server}

\item{session}{session server is operating on}
}
\description{
Server code for the gadget.
}
