% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_lab.R
\name{subset_lab}
\alias{subset_lab}
\title{Subset data and save attributes}
\usage{
subset_lab(x, ...)
}
\arguments{
\item{x}{A data frame or matrix}

\item{subset}{A logical vector}

\item{select}{A character vector with column names}
}
\value{
The subsetted data set
}
\description{
Subset data and save attributes
}
\examples{
subset(airquality, Temp > 80, select = c(Ozone, Temp))
subset(airquality, Day == 1, select = -Temp)
subset(airquality, select = Ozone:Wind)

with(airquality, subset(Ozone, Temp > 80))

## sometimes requiring a logical 'subset' argument is a nuisance
nm <- rownames(state.x77)
start_with_M <- nm \%in\% grep("^M", nm, value = TRUE)
subset(state.x77, start_with_M, Illiteracy:Murder)
# but in recent versions of R this can simply be
subset(state.x77, grepl("^M", nm), Illiteracy:Murder)

}
