% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_replace.R
\name{View}
\alias{View}
\title{View}
\usage{
View(datain, replace = T)
}
\arguments{
\item{datain}{a data frame or list of data frames}

\item{replace}{Set equal to false to use normal view}
}
\value{
Either the View, or the app View from viewenhanceAddin
}
\description{
Replaces View as given by R Studio.
This automatically opens the viewer app.
Disable this by setting replace = F
}
