% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server_in}
\alias{server_in}
\title{Server code for the gadget.}
\usage{
server_in(datain, list_true)
}
\arguments{
\item{datain}{data entered}

\item{list_true}{whether list_true is true or not}
}
\description{
Server code for the gadget.
}
