% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_lab.R
\name{subset_lab}
\alias{subset_lab}
\title{Subset data and save attributes.}
\usage{
subset_lab(x, ..., col_lim = 1000)
}
\arguments{
\item{x}{A data frame or matrix.}

\item{...}{Other arguments passed to subset. In particular, we give subset
and select arguments here (see examples and subset help for more)}

\item{col_lim}{A limit on the number of columns the attributes are applied to.}
}
\value{
The subsetted data set
}
\description{
Note that this function is only designed to handle data frames
Note that for large data sets this will take a long time,
so by default this is only reapplied for the first 1000 columns
}
\examples{
subset_lab(airquality, Temp > 80, select = c(Ozone, Temp))
subset_lab(airquality, Day == 1, select = -Temp)
subset_lab(airquality, select = Ozone:Wind)

## sometimes requiring a logical 'subset' argument is a nuisance
nm <- rownames(state.x77)
start_with_M <- nm \%in\% grep("^M", nm, value = TRUE)
subset_lab(state.x77, start_with_M, Illiteracy:Murder)
# but in recent versions of R this can simply be
subset_lab(state.x77, grepl("^M", nm), Illiteracy:Murder)

}
