% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewenhanceAddin.R
\name{viewenhanceAddin}
\alias{viewenhanceAddin}
\title{Shiny view of data}
\usage{
viewenhanceAddin()
}
\value{
A shiny box, which, when options are chosen, will put a View command into the console
}
\description{
Shiny view of data
}
